+++ index.js
@@ -298,7 +298,7 @@ var _toString2 = _interopRequireDefault(_toString);
 
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
-var version = '10.7.1';
+var version = '10.8.0';
 
 var validator = {
   version: version,