+++ validator.js
@@ -1722,7 +1722,7 @@ function normalizeEmail(email, options) {
   return parts.join('@');
 }
 
-var version = '10.7.1';
+var version = '10.8.0';
 
 var validator = {
   version: version,