+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Added `isIdentityCard()`
   ([#846](https://github.com/chriso/validator.js/pull/846))
+- Better error when validators are passed an invalid type
+  ([#895](https://github.com/chriso/validator.js/pull/895))
 - Locales are now exported
   ([#890](https://github.com/chriso/validator.js/pull/890),
    [#892](https://github.com/chriso/validator.js/pull/892))